PR-G cleanups: AGENTS.md compliance + convention drift (#206)
A — Fix getattr(obj.db, ...) violation in wounds/constants.py:54;
use direct attribute access (returns None when unset) per AGENTS.md.

B — Rewrite stale ORGAN_DISPLAY header comment referencing the
deleted Organ.return_appearance to describe the Evennia-standard
db.desc flow populated by Organ.configure_from_harvest.

C — Replace imperative species init in Character.at_object_creation
with declarative AttributeProperty("human", category="identity"),
matching the convention used by other identity attributes.

D — Relocate ORGAN_DISPLAY + helpers from world.medical.constants
to new world.anatomy.organs module (mirrors world.anatomy.severed_parts
shape).  Re-exported from world.anatomy.  Imports updated in
typeclasses.items and world.tests.test_organ_display.  Clean break:
no backward-compat re-export from medical.constants.

E — Replace try/except AttributeError with hasattr() check around
corpse.get_decay_stage() in Appendage.configure_from_sever; hasattr
is the AGENTS.md-sanctioned idiom for typeclass method existence.

All 1152 tests pass.
